move create plate data to data class in gen5
